Annual agricultural production data of Peru

The cropdatape package is based on the Integrated System of Agricultural Statistics (SIEA) of Peruvian Agriculture Ministry. The variables are: sowing and harvest area, production, crop yield, price per plot by region. The range of year covers from 2004 to 2014. In case of 2004, it just considered from August to December (mid-seasson). Some missing values (NA) indicate that in some periods do not exists agricultural activity.

cropdatape

Format

Data frame with --- rows and 9 columns.

crop

factor with 6 levels

department

Peruvian departments. Factor with 24 levels

year

Numeric. Range of years from 2004 to 2014

month

Factor with 12 levels.

sowa

Sowing area measured by hectares (ha)

harva

Harvest area measured by hectare (ha)

production

Production of crop measured by tons (t)

yield

Yield of crop measured by kilograms per hectare (kg/ha)

pricePlot

Price per plot. The peruvian currency peruvian nuevo sol

Source

http://siea.minagri.gob.pe/siea/?q=publicaciones/anuarios-estadisticos

Examples

cropdatape

#> # A tibble: 15,080 x 9
#>    crop  department  year month      sowa harva production yield pricePlot
#>    <fct> <fct>      <dbl> <fct>     <dbl> <dbl>      <dbl> <dbl>     <dbl>
#>  1 rice  Amazonas    2004 August      949  2855     20133. 7052.     0.631
#>  2 rice  Amazonas    2004 December   3393  1763     12183. 6910.     0.514
#>  3 rice  Amazonas    2004 November   2040  2714     18867  6952.     0.617
#>  4 rice  Amazonas    2004 October    1491  2812     19760. 7027.     0.640
#>  5 rice  Amazonas    2004 September  1059  2544     17964  7061.     0.636
#>  6 rice  Ancash      2004 August       NA    NA        NA    NA     NA    
#>  7 rice  Ancash      2004 December    350    NA        NA    NA     NA    
#>  8 rice  Ancash      2004 November    984    NA        NA    NA     NA    
#>  9 rice  Ancash      2004 October     782    NA        NA    NA     NA    
#> 10 rice  Ancash      2004 September    15    NA        NA    NA     NA    
#> # ... with 15,070 more rows